move spec to base, use new rebase-helper api
