+++ tests/utils.py
@@ -1,5 +1,8 @@
+import inspect
 import os
 from contextlib import contextmanager
+
+from rebasehelper.specfile import SpecFile
 
 
 @contextmanager
@@ -18,3 +21,11 @@ def cwd(target):
         yield
     finally:
         os.chdir(curdir)
+
+
+def get_specfile(path: str) -> SpecFile:
+    s = inspect.signature(SpecFile)
+    if "changelog_entry" in s.parameters:
+        return SpecFile(path=path, changelog_entry=None)
+    else:
+        return SpecFile(path=path)
